+++ packages/mobile/src/screens/auth/FacebookButton.tsx
@@ -18,7 +18,7 @@ interface Props {
   label: string;
 }
 
-export const FacebookButton: React.FC<Props> = React.memo(({ label }) => {
+export const FacebookButton: React.FC<Props> = ({ label }) => {
   const { service, loading } = useAuth();
   const { navigate } = useNavigation<AuthStackNavProp>();
   const [fbPressed, setFbPressed] = useState(false);
@@ -63,6 +63,4 @@ export const FacebookButton: React.FC<Props> = React.memo(({ label }) => {
       {paddedLabel}
     </Button>
   );
-});
-
-FacebookButton.displayName = 'FacebookButton';
+};