+++ services/backend/src/auth/facebook/__tests__/mobile.ts
@@ -1,3 +1,11 @@
+import { AuthBody, SignInBody } from '@whitewater-guide/commons';
+import Koa from 'koa';
+import get from 'lodash/get';
+import { Profile } from 'passport-facebook';
+import FacebookTokenStrategy from 'passport-facebook-token';
+import superagent from 'superagent';
+import { SuperTest, Test } from 'supertest';
+import agent from 'supertest-koa-agent';
 import db, { holdTransaction, rollbackTransaction } from '~/db';
 import { redis } from '~/redis';
 import {
@@ -8,16 +16,8 @@ import {
   NEW_FB_PROFILE_W_LOCALE,
 } from '~/seeds/test/01_users';
 import { countRows, UUID_REGEX } from '~/test';
-import { AuthBody, SignInBody } from '@whitewater-guide/commons';
-import Koa from 'koa';
-import get from 'lodash/get';
-import { Profile } from 'passport-facebook';
-import FacebookTokenStrategy from 'passport-facebook-token';
-import superagent from 'superagent';
-import { SuperTest, Test } from 'supertest';
-import agent from 'supertest-koa-agent';
 import { createApp } from '../../../app';
-import { MailType, sendMail } from '../../mail';
+import { sendWelcome } from '../../mail';
 
 jest.mock('../../mail');
 
@@ -145,19 +145,14 @@ describe.each([
   });
 
   it('should send welcome email', async () => {
-    const email = get(mockProfile, 'emails.0.value');
-    if (email) {
-      expect(sendMail).toHaveBeenCalledWith(MailType.WELCOME_VERIFIED, email, {
-        user: {
-          id: expect.stringMatching(UUID_REGEX),
-          name: `${mockProfile.name!.givenName} ${
-            mockProfile.name!.familyName
-          }`,
-        },
-      });
-    } else {
-      expect(sendMail).not.toHaveBeenCalled();
-    }
+    const email = get(mockProfile, 'emails.0.value', null);
+    expect(sendWelcome).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: expect.stringMatching(UUID_REGEX),
+        name: `${mockProfile.name!.givenName} ${mockProfile.name!.familyName}`,
+        email,
+      }),
+    );
   });
 });
 
@@ -283,7 +278,7 @@ describe('existing user', () => {
   });
 
   it('should not send welcome email', async () => {
-    expect(sendMail).not.toHaveBeenCalled();
+    expect(sendWelcome).not.toHaveBeenCalled();
   });
 });