+++ services/backend/src/auth/facebook/logger.ts
@@ -1,5 +1,0 @@
-import authLogger from '../logger';
-
-const logger = authLogger.child({ strategy: 'facebook' });
-
-export default logger;