+++ services/backend/src/auth/mail/index.ts
@@ -1,2 +1,3 @@
 export * from './sendMail';
+export * from './sendWelcome';
 export * from './types';