+++ services/web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.1.433-apple.1](https://github.com/doomsower/whitewater/compare/584b3252b25408c4b54af5f4df4134d9c699554e...c9713f7f65dfe49f913b52b04e6a2d67ecfb1420) (2020-07-27)
+
+## [0.1.433-apple.1](https://github.com/doomsower/whitewater/compare/2ac31c235c901f5383b50c5af578559efd368c9e...584b3252b25408c4b54af5f4df4134d9c699554e) (2020-07-27)
+
+## [0.1.433-apple.1](https://github.com/doomsower/whitewater/compare/c9c1e2536d82881f7dcb91021670cbc53e8662d9...2ac31c235c901f5383b50c5af578559efd368c9e) (2020-07-26)
+
+## [0.1.433-apple.1](https://github.com/doomsower/whitewater/compare/234a0f84e82ecd434438516b9f68a1e1726b2852...c9c1e2536d82881f7dcb91021670cbc53e8662d9) (2020-07-26)
+
+## [0.1.433-apple.1](https://github.com/doomsower/whitewater/compare/d1106419e3007e266d94a284c12b4cd3984b2266...234a0f84e82ecd434438516b9f68a1e1726b2852) (2020-07-26)
+
+## [0.1.433-apple.0](https://github.com/doomsower/whitewater/compare/f7720ae627f6e5e276a64c52d8a76004af13fe64...d1106419e3007e266d94a284c12b4cd3984b2266) (2020-07-26)
+
 ## [0.1.432](https://github.com/doomsower/whitewater/compare/69fdd1af80a4b22b4088243c18eb181cf7faeab0...f7720ae627f6e5e276a64c52d8a76004af13fe64) (2020-07-26)
 
 ## [0.1.431](https://github.com/doomsower/whitewater/compare/5d99a781cf33884d1fa71ced28978f3b8915d104...69fdd1af80a4b22b4088243c18eb181cf7faeab0) (2020-07-26)